+++ app/src/main/kotlin/com/blocksdecoded/coinwave/data/coins/remote/ICoinClient.kt
@@ -6,7 +6,7 @@ import com.blocksdecoded.coinwave.data.model.CoinsResponse
 import io.reactivex.Single
 
 interface ICoinClient {
-    fun getCoins(pageSize: Int): Single<CoinsResponse>
+    fun getCoins(): Single<CoinsResponse>
 
-    fun getHistory(chartName: String, period: ChartPeriodEnum): Single<HistoryResponse>
+    fun getHistory(coin: String, period: ChartPeriodEnum): Single<HistoryResponse>
 }