+++ app/src/main/kotlin/com/blocksdecoded/coinwave/data/coins/remote/ICoinClientConfig.kt
@@ -1,9 +1,10 @@
 package com.blocksdecoded.coinwave.data.coins.remote
 
+import okhttp3.HttpUrl
+
 interface ICoinClientConfig {
-    var coinUrl: String
-    val ipnsKey: String
-        get() = coinUrl
-            .substringAfter("ipns", "QmT3qT84rdZ9tkUssYLNkKbPcFTkutFy2S8uizvKWQbTMg")
-            .replace("/", "")
+    var ipnsUrl: String
+
+    val ipnsPath: String
+        get() = HttpUrl.get(ipnsUrl).encodedPath()
 }