+++ app/src/main/kotlin/com/blocksdecoded/coinwave/data/config/ConfigProvider.kt
@@ -1,5 +1,5 @@
 package com.blocksdecoded.coinwave.data.config
 
 data class ConfigProvider(
-    override var coinUrl: String
+    override var ipnsUrl: String
 ) : IConfigProvider