+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ###Fixed
 - Watchlist refresh
+- Bootstrap config
 
 ## [1.0.0 (12)] - 2019-03-22
 ###Added